+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Products.MeetingLiege Changelog
 ===============================
 
-4.1rc8 (unreleased)
+4.1rc8 (2020-04-29)
 -------------------
 
 - Removed custom inserting method `on_decision_first_word`, now using the default `on_item_decision_first_words` that is doing the same